refactor: 웹소켓 메시지 패키지 변경
