--- src/main/java/com/game/contraband/infrastructure/message/ExceptionCode.java
@@ -1,5 +1,0 @@
-package com.game.contraband.infrastructure.message;
-
-public enum ExceptionCode {
-    UNKNOWN_ERROR
-}
