+++ src/main/java/com/game/contraband/infrastructure/websocket/message/ExceptionCode.java
@@ -1,0 +1,5 @@
+package com.game.contraband.infrastructure.websocket.message;
+
+public enum ExceptionCode {
+    UNKNOWN_ERROR
+}
